+++ adminer/drivers/pgsql.inc.php
@@ -853,8 +853,9 @@ ORDER BY ordinal_position');
 	function routines() {
 		return get_rows('SELECT specific_name AS "SPECIFIC_NAME", routine_type AS "ROUTINE_TYPE", routine_name AS "ROUTINE_NAME", type_udt_name AS "DTD_IDENTIFIER"
 FROM information_schema.routines
-WHERE routine_schema = current_schema()
-ORDER BY SPECIFIC_NAME');
+WHERE routine_schema = current_schema()' . (connection()->flavor == 'cockroach' ? '' : "
+AND substring(specific_name, '[0-9]+\$')::oid NOT IN (SELECT objid FROM pg_catalog.pg_depend WHERE classid = 'pg_proc'::regclass AND deptype = 'e')") . '
+ORDER BY SPECIFIC_NAME'); // 'e' - functions created by extensions
 	}
 
 	function routine_languages() {